[$] Reto 4 - Variante 1: Ajustes finales clase Trabajador & Pruebas Locales

+++ src/main/java/co/edu/udea/jcarlosj/reto_4/Nomina.java
@@ -63,7 +63,7 @@ public class Nomina {
         for( Trabajador trabajador : this .trabajadores )
             sumatoriaEdades += trabajador. calcularEdad();
 
-        return sumatoriaEdades / this .trabajadores .size();
+        return (double) sumatoriaEdades / this .trabajadores .size();
     }
 
     public double desviacionEstandarEdadNomina() {
